refactor(stressgres): extract transient-fault tolerance into resilience.rs
The transient-fault classification and retry helpers had no coupling to the rest
of runner.rs (they only touch postgres::Error / anyhow / std), so move them into
a dedicated `resilience` module. runner.rs now just imports `tolerate_transient`;
everything else (needle list, is_transient_* classifiers, backoff) is private to
the module.

Pure code move — no behavior change — plus unit tests for the classifier's
string-matching core, which was previously untestable buried in runner.rs.
